+++ src/server.rs
@@ -35,7 +35,7 @@ pub fn config_server(
     let mut server_config = ServerConfig::with_crypto(Arc::new(server_crypto));
 
     Arc::get_mut(&mut server_config.transport)
-        .unwrap()
+        .expect("safevalue")
         .max_concurrent_uni_streams(0_u8.into());
 
     Ok(server_config)
@@ -55,7 +55,7 @@ pub fn certificate_info(connection: &Connection) -> Result<String> {
                             .iter_common_name()
                             .next()
                             .and_then(|cn| cn.as_str().ok())
-                            .expect("the issuer of the certificate is not valid");
+                            .context("the issuer of the certificate is not valid")?;
                         info!("Connected Client Name : {}", issuer);
                         return Ok(String::from(issuer));
                     }
@@ -83,35 +83,35 @@ pub async fn server_handshake(
     let mut version_buf = Vec::new();
     version_buf.resize(len.try_into()?, 0);
     recv.read_exact(version_buf.as_mut_slice()).await?;
-    let version = String::from_utf8(version_buf).unwrap();
+    let version = String::from_utf8(version_buf).context("string from utf")?;
 
     match min_version.cmp(&version) {
         Less | Equal => match max_version.cmp(&version) {
             Greater => {
-                send.write_all(&handshake_buffer(Some(env!("CARGO_PKG_VERSION"))))
+                send.write_all(&handshake_buffer(Some(env!("CARGO_PKG_VERSION")))?)
                     .await?;
                 info!("Compatible Version");
             }
             Less | Equal => {
-                send.write_all(&handshake_buffer(None)).await?;
+                send.write_all(&handshake_buffer(None)?).await?;
                 bail!("Incompatible version")
             }
         },
         Greater => {
-            send.write_all(&handshake_buffer(None)).await?;
+            send.write_all(&handshake_buffer(None)?).await?;
             bail!("Incompatible version")
         }
     }
     Ok(())
 }
 
-fn handshake_buffer(resp: Option<&str>) -> Vec<u8> {
-    let resp_data = bincode::serialize::<Option<&str>>(&resp).unwrap();
+fn handshake_buffer(resp: Option<&str>) -> Result<Vec<u8>> {
+    let resp_data = bincode::serialize::<Option<&str>>(&resp)?;
     let resp_data_len = u64::try_from(resp_data.len())
-        .expect("less than u64::MAX")
+        .context("less than u64::MAX")?
         .to_le_bytes();
     let mut resp_buf = Vec::with_capacity(resp_data_len.len() + resp_data.len());
     resp_buf.extend(resp_data_len);
     resp_buf.extend(resp_data);
-    resp_buf
+    Ok(resp_buf)
 }